refactor: get sample app compiling with new task runner
commit-id:6b8dfde6
