+++ Example/Wendy/AppDelegate.swift
@@ -9,7 +9,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // Override point for customization after application launch.
         window = UIWindow(frame: UIScreen.main.bounds)
 
-        Wendy.setup(tasksFactory: ExampleAppPendingTasksFactory())
+        Wendy.setup(taskRunner: MyWendyTaskRunner())
         #if DEBUG
         WendyConfig.debug = true
         #endif
