+++ Example/Wendy/Type/TaskTags.swift
@@ -1,0 +1,13 @@
+//
+//  TaskTags.swift
+//  Wendy_Example
+//
+//  Created by Levi Bostian on 1/14/24.
+//  Copyright © 2024 CocoaPods. All rights reserved.
+//
+
+import Foundation
+
+enum TaskTag: String {
+    case addGroceryListItem
+}
